First implementation of UTF-8 path support for Zip reader.
This uses the existing "update_pathname_utf8" logic to
accept the UTF8 pathname.  The tests verify that this
supports UTF-8 paths stored using GP#11 or 0x7075 extension.
More testing is certainly needed...

Note:  A lot of the diff here is reshuffling of an internal API
so it can accept non-null-terminated strings.